--- a/lib/spack/spack/cmd/develop.py
+++ b/lib/spack/spack/cmd/develop.py
@@ -9,7 +9,9 @@ import llnl.util.tty as tty
 
 import spack.cmd
 import spack.cmd.common.arguments as arguments
+import spack.spec
 import spack.util.path
+import spack.version
 from spack.error import SpackError
 
 description = "add a spec to an environment's dev-build information"
@@ -61,7 +63,7 @@ def develop(parser, args):
                 tty.msg(msg)
                 continue
 
-            spec = spack.spec.Spec(entry["spec"])
+            spec = spack.spec.concrete_spec_from_old_syntax(entry["spec"])
             pkg_cls = spack.repo.path.get_pkg_class(spec.name)
             pkg_cls(spec).stage.steal_source(abspath)
 
@@ -75,8 +77,11 @@ def develop(parser, args):
         raise SpackError("spack develop requires at most one named spec")
 
     spec = specs[0]
-    if not spec.versions.concrete:
+    version = spec.versions.concrete_range_as_version
+    if not version:
         raise SpackError("Packages to develop must have a concrete version")
+
+    spec.versions = spack.version.VersionList([version])
 
     # default path is relative path to spec.name
     path = args.path or spec.name