--- a/lib/spack/spack/compilers/apple_clang.py
+++ b/lib/spack/spack/compilers/apple_clang.py
@@ -41,7 +41,7 @@ class AppleClang(spack.compilers.clang.Clang):
     @property
     def cxx11_flag(self):
         # Spack's AppleClang detection only valid from Xcode >= 4.6
-        if self.real_version < spack.version.ver("4.0"):
+        if self.real_version < spack.version.ver("=4.0"):
             raise spack.compiler.UnsupportedCompilerFlag(
                 self, "the C++11 standard", "cxx11_flag", "Xcode < 4.0"
             )
@@ -49,38 +49,38 @@ class AppleClang(spack.compilers.clang.Clang):
 
     @property
     def cxx14_flag(self):
-        if self.real_version < spack.version.ver("5.1"):
+        if self.real_version < spack.version.ver("=5.1"):
             raise spack.compiler.UnsupportedCompilerFlag(
                 self, "the C++14 standard", "cxx14_flag", "Xcode < 5.1"
             )
-        elif self.real_version < spack.version.ver("6.1"):
+        elif self.real_version < spack.version.ver("=6.1"):
             return "-std=c++1y"
 
         return "-std=c++14"
 
     @property
     def cxx17_flag(self):
-        if self.real_version < spack.version.ver("6.1"):
+        if self.real_version < spack.version.ver("=6.1"):
             raise spack.compiler.UnsupportedCompilerFlag(
                 self, "the C++17 standard", "cxx17_flag", "Xcode < 6.1"
             )
-        elif self.real_version < spack.version.ver("10.0"):
+        elif self.real_version < spack.version.ver("=10.0"):
             return "-std=c++1z"
         return "-std=c++17"
 
     @property
     def cxx20_flag(self):
-        if self.real_version < spack.version.ver("10.0"):
+        if self.real_version < spack.version.ver("=10.0"):
             raise spack.compiler.UnsupportedCompilerFlag(
                 self, "the C++20 standard", "cxx20_flag", "Xcode < 10.0"
             )
-        elif self.real_version < spack.version.ver("13.0"):
+        elif self.real_version < spack.version.ver("=13.0"):
             return "-std=c++2a"
         return "-std=c++20"
 
     @property
     def cxx23_flag(self):
-        if self.real_version < spack.version.ver("13.0"):
+        if self.real_version < spack.version.ver("=13.0"):
             raise spack.compiler.UnsupportedCompilerFlag(
                 self, "the C++23 standard", "cxx23_flag", "Xcode < 13.0"
             )
@@ -90,7 +90,7 @@ class AppleClang(spack.compilers.clang.Clang):
 
     @property
     def c99_flag(self):
-        if self.real_version < spack.version.ver("4.0"):
+        if self.real_version < spack.version.ver("=4.0"):
             raise spack.compiler.UnsupportedCompilerFlag(
                 self, "the C99 standard", "c99_flag", "< 4.0"
             )
@@ -98,7 +98,7 @@ class AppleClang(spack.compilers.clang.Clang):
 
     @property
     def c11_flag(self):
-        if self.real_version < spack.version.ver("4.0"):
+        if self.real_version < spack.version.ver("=4.0"):
             raise spack.compiler.UnsupportedCompilerFlag(
                 self, "the C11 standard", "c11_flag", "< 4.0"
             )
@@ -106,7 +106,7 @@ class AppleClang(spack.compilers.clang.Clang):
 
     @property
     def c17_flag(self):
-        if self.real_version < spack.version.ver("11.0"):
+        if self.real_version < spack.version.ver("=11.0"):
             raise spack.compiler.UnsupportedCompilerFlag(
                 self, "the C17 standard", "c17_flag", "< 11.0"
             )
@@ -114,7 +114,7 @@ class AppleClang(spack.compilers.clang.Clang):
 
     @property
     def c23_flag(self):
-        if self.real_version < spack.version.ver("11.0.3"):
+        if self.real_version < spack.version.ver("=11.0.3"):
             raise spack.compiler.UnsupportedCompilerFlag(
                 self, "the C23 standard", "c23_flag", "< 11.0.3"
             )