--- a/lib/spack/spack/compilers/cce.py
+++ b/lib/spack/spack/compilers/cce.py
@@ -5,7 +5,7 @@
 import os
 
 from spack.compiler import Compiler, UnsupportedCompilerFlag
-from spack.version import ver
+from spack.version import Version
 
 
 class Cce(Compiler):
@@ -58,7 +58,7 @@ class Cce(Compiler):
     @property
     def is_clang_based(self):
         version = self._real_version or self.version
-        return version >= ver("=9.0") and "classic" not in str(version)
+        return version >= Version("9.0") and "classic" not in str(version)
 
     version_argument = "--version"
     version_regex = r"[Cc]ray (?:clang|C :|C\+\+ :|Fortran :) [Vv]ersion.*?(\d+(\.\d+)+)"
@@ -98,9 +98,9 @@ class Cce(Compiler):
     def c99_flag(self):
         if self.is_clang_based:
             return "-std=c99"
-        elif self.real_version >= ver("=8.4"):
+        elif self.real_version >= Version("8.4"):
             return "-h std=c99,noconform,gnu"
-        elif self.real_version >= ver("=8.1"):
+        elif self.real_version >= Version("8.1"):
             return "-h c99,noconform,gnu"
         raise UnsupportedCompilerFlag(self, "the C99 standard", "c99_flag", "< 8.1")
 
@@ -108,7 +108,7 @@ class Cce(Compiler):
     def c11_flag(self):
         if self.is_clang_based:
             return "-std=c11"
-        elif self.real_version >= ver("=8.5"):
+        elif self.real_version >= Version("8.5"):
             return "-h std=c11,noconform,gnu"
         raise UnsupportedCompilerFlag(self, "the C11 standard", "c11_flag", "< 8.5")