--- a/lib/spack/spack/compilers/clang.py
+++ b/lib/spack/spack/compilers/clang.py
@@ -100,24 +100,24 @@ class Clang(Compiler):
 
     @property
     def cxx11_flag(self):
-        if self.real_version < ver("3.3"):
+        if self.real_version < ver("=3.3"):
             raise UnsupportedCompilerFlag(self, "the C++11 standard", "cxx11_flag", "< 3.3")
         return "-std=c++11"
 
     @property
     def cxx14_flag(self):
-        if self.real_version < ver("3.4"):
+        if self.real_version < ver("=3.4"):
             raise UnsupportedCompilerFlag(self, "the C++14 standard", "cxx14_flag", "< 3.5")
-        elif self.real_version < ver("3.5"):
+        elif self.real_version < ver("=3.5"):
             return "-std=c++1y"
 
         return "-std=c++14"
 
     @property
     def cxx17_flag(self):
-        if self.real_version < ver("3.5"):
+        if self.real_version < ver("=3.5"):
             raise UnsupportedCompilerFlag(self, "the C++17 standard", "cxx17_flag", "< 3.5")
-        elif self.real_version < ver("5.0"):
+        elif self.real_version < ver("=5.0"):
             return "-std=c++1z"
 
         return "-std=c++17"
@@ -128,21 +128,21 @@ class Clang(Compiler):
 
     @property
     def c11_flag(self):
-        if self.real_version < ver("3.0"):
+        if self.real_version < ver("=3.0"):
             raise UnsupportedCompilerFlag(self, "the C11 standard", "c11_flag", "< 3.0")
-        if self.real_version < ver("3.1"):
+        if self.real_version < ver("=3.1"):
             return "-std=c1x"
         return "-std=c11"
 
     @property
     def c17_flag(self):
-        if self.real_version < ver("6.0"):
+        if self.real_version < ver("=6.0"):
             raise UnsupportedCompilerFlag(self, "the C17 standard", "c17_flag", "< 6.0")
         return "-std=c17"
 
     @property
     def c23_flag(self):
-        if self.real_version < ver("9.0"):
+        if self.real_version < ver("=9.0"):
             raise UnsupportedCompilerFlag(self, "the C23 standard", "c23_flag", "< 9.0")
         return "-std=c2x"