… lib/spack/docs/basic_usage.rst
… lib/spack/docs/packaging_guide.rst
… lib/spack/spack/compilers/gcc.py
… lib/spack/spack/compilers/intel.py
… lib/spack/spack/compilers/xl.py
… lib/spack/spack/concretize.py
… lib/spack/spack/cray_manifest.py
… lib/spack/spack/database.py
… lib/spack/spack/directives.py
… lib/spack/spack/environment/environment.py
… lib/spack/spack/fetch_strategy.py
… lib/spack/spack/installer.py
… lib/spack/spack/package_base.py
… lib/spack/spack/package_prefs.py
… lib/spack/spack/parser.py
… lib/spack/spack/schema/packages.py
… lib/spack/spack/solver/asp.py
… lib/spack/spack/spec.py
… lib/spack/spack/spec_list.py
… lib/spack/spack/test/architecture.py
… lib/spack/spack/test/cmd/ci.py
… lib/spack/spack/test/cmd/config.py
… lib/spack/spack/test/cmd/dev_build.py
… lib/spack/spack/test/cmd/develop.py
… lib/spack/spack/test/cmd/env.py
… lib/spack/spack/test/cmd/fetch.py
… lib/spack/spack/test/cmd/install.py
… lib/spack/spack/test/cmd/spec.py
… lib/spack/spack/test/cmd/stage.py
… lib/spack/spack/test/compilers/basics.py