@haampie 2023-03-22
```
PKG-DIRECTIVES: 3 issues found
1. py-apache-beam: dependency on py-dill@0.3.1.1:0.3.1 cannot be satisfied by known versions of py-dill
    happening in /home/runner/work/spack/spack/var/spack/repos/builtin/packages/py-apache-beam/package.py
    known versions of py-dill are 0.3.6, 0.3.5.1, 0.3.4, 0.3.1, 0.2.9, 0.2.7, 0.2.6, 0.2.5, 0.2.4, 0.2.3, 0.2.2, 0.2.1, 0.2
2. py-cutadapt: dependency on py-dnaio@0.3.0:0.3 cannot be satisfied by known versions of py-dnaio
    happening in /home/runner/work/spack/spack/var/spack/repos/builtin/packages/py-cutadapt/package.py
    known versions of py-dnaio are 0.9.1, 0.4.2, 0.3
3. py-sphinx-tabs: dependency on py-docutils@0.16.0:0.16 cannot be satisfied by known versions of py-docutils
    happening in /home/runner/work/spack/spack/var/spack/repos/builtin/packages/py-sphinx-tabs/package.py
    known versions of py-docutils are 0.19, 0.18.1, 0.18, 0.17.1, 0.17, 0.16, 0.15.2, 0.14, 0.13.1, 0.12
```

looks correct to me