@alalazo 2023-04-04
So far this is improving the setup phase in a noticeable way:
[radiuss_versions.csv](https://github.com/spack/spack/files/11149331/radiuss_versions.csv) (e424fd49ad9dfd9926f21927f9e13a6ee5633973)
[radiuss_develop.csv](https://github.com/spack/spack/files/11149332/radiuss_develop.csv) (6a2f80e2c655cba38a4efbda862a7d7060291444)
[radiuss.txt](https://github.com/spack/spack/files/11149333/radiuss.txt)

Here's the plot:
![totals](https://user-images.githubusercontent.com/4199709/229817462-dbf96a2f-45c9-4595-988e-227f39b1b61a.png)