@trws 2023-04-04
Initial parse might be a bit slower, translating to ClosedOpenRange rather than a closed/closed?  Or maybe there are a couple more string passes from checking for `=` in some places? I would expect the overall to be better, and @haampie and I were just discussing and I think going the extra step to make StandardVersion itself a ClosedOpenRange to get the full comparison perf boost should remain feasible after this is in, maybe also GitVersion but they're strange enough that's not going to entirely go away.