@tgamblin 2023-04-05
> Exact matching logic: on develop the behavior of spack install pkg@3.1 would install an exact match of pkg@=3.1 if both version 3.1 and 3.1.1 were defined in the package. For backwards compatibility this behavior is preserved. In the future we could change this behavior since users can now distinguish between pkg@=3.1 and pkg@3.1.

I actually think this should prefer 3.1.1 -- any reason not to do that?  I don't think we guarantee backward compatibility at this level, but if there's a nasty consequence to this that I'm missing lemme know.