@tgamblin 2023-04-05
Ok, a nasty consequence for anyone with unresolved git versions installed: I happen to have `zlib@04f42ceca40f73e2978b50e93806c2a18c1281fc` installed. 

On `develop`:
```console
(spackbook):spack> spack find zlib
-- darwin-monterey-m1 / apple-clang@14.0.0 ----------------------
zlib@1.2.13  zlib@1.2.13  zlib@04f42ceca40f73e2978b50e93806c2a18c1281fc
```

On this branch:
```console
(spackbook):spack> spack find zlib
==> Error: git ref '04f42ceca40f73e2978b50e93806c2a18c1281fc' cannot be looked up: call attach_git_lookup_from_package first
```

Is there some way we can grandfather in installations like this?  I think it would be ok if we treated them like `@git.ref=develop` to avoid an extra lookup.