@haampie 2023-04-05
Okay, two things: 

1. `from_node_dict` was another instance where the git repo was simply not attached after constructing a Spec, which is now fixed, so specs with just shas in the db can now resolve to their spack versions
2. whenever the database contains a spec with git sha version w/o associated spack version, that's unresolvable due to package.py dropping the git attribute or so, spack now issues a warning and recommends to remove it. the git versions gets replaced by a normal version; this is what it's like on develop right now, and the reason to not assign `=develop` is because it mutates the immutable spec and would lead to weird issues in reuse concretization.