@tgamblin 2023-04-05
I removed my zlib version earlier but I still have a one-off `clingo` commit in the DB.  I get this when doing `spack find` now:

```console
(spackbook):spack> spack find
fatal: repository 'foo' does not exist
==> Error: Invalid record in Spack database: hash: ebfvstdloypshiz37hhigps7pct3nzit, cause: ProcessError: Command exited with status 128:
    '/Users/gamblin2/src/spack/var/spack/environments/default/.spack-env/view/bin/git' '-c' 'advice.detachedHead=false' 'clone' '--bare' '--quiet' 'foo' '/Users/gamblin2/Workspace/src/spack/lib/spack/spack/foo'
/Users/gamblin2/Workspace/src/spack/opt/spack/.spack-db/index.json
```