@tgamblin 2023-04-05
Ok so part of this may be outside the scope of this particular PR, but it's all nasty and probably needs fixing.

More on my `zlib@commit`: I was playing wth git versions and added a `git` property to `zlib`, then installed `zlib@37ed2112a1727293f4fd431df85dc821ee475d3d`.  I've since *removed* the git URL for zlib -- I didn't bother upstreaming it though I do think we should add git URLs to more things.

Since my last comment above, I've *removed* that `zlib` installation, but it's still in the DB:

```console
(spackbook):spack> spack find zlib
-- darwin-monterey-m1 / apple-clang@14.0.0 ----------------------
zlib@1.2.13  zlib@1.2.13

-- darwin-ventura-m1 / apple-clang@14.0.0 -----------------------
zlib@1.2.13
==> 3 installed packages
```

```console
(spackbook):spack> spack find -m zlib
-- darwin-bigsur-skylake / apple-clang@13.0.0 -------------------
zlib@37ed2112a1727293f4fd431df85dc821ee475d3d

-- darwin-monterey-m1 / apple-clang@14.0.0 ----------------------
zlib@1.2.13  zlib@1.2.13

-- darwin-ventura-m1 / apple-clang@14.0.0 -----------------------
zlib@1.2.13
==> 4 installed packages
```

On this branch I can't run any find command:
```console
(spackbook):spack> spack find
==> Error: Invalid record in Spack database: hash: ebfvstdloypshiz37hhigps7pct3nzit, cause: AttributeError: type object 'Zlib' has no attribute 'git'
/Users/gamblin2/Workspace/src/spack/opt/spack/.spack-db/index.json
```

It looks like it's taking the unresolved git commit out of the DB and attempting to resolve it.  Basically I think that shouldn't ever happen -- but it seemed like that's what you thought too:

> whenever the database contains a spec with unresolvable git sha version w/o associated spack version (due to package.py dropping the git attribute or so), spack now issues a warning and recommends to remove it.

Is it attempting to *resolve* the SHA?  I think it shouldn't even do that.

I don't really agree with this:

> the reason to not assign `=develop` is because it mutates the otherwise immutable spec, and would lead to weird issues in reuse concretization/hashing. 

Concrete hashes in Spack are assigned at concretization time and are respected forever.  So you can still respect the identifier for the installation (the dag hash) while changing how a version on that spec is compared.  It should never mutate the hash, and in reuse concretization the hashes are independent of the spec contents -- you *can* tell the solver to treat the version like it's `=develop` without changing the hash or altering reuse.  Every hash spec in reuse currently implies its dependencies (by hash) so you won't run into any issues with consistency in the sub-DAG rooted at the unresolved-git-version spec.  It'll only affect how constraints from *dependents* are evaluated.

So I still think the thing to do is to read the git commit version from the database, keep it around, and have it compare as `=develop` or `=inf` or something.  I guess the question is where that transformation should happen.  I *think* it's fine to do on read, though it would be nice if the spec serialized to the same JSON as before, but it *necessarily* doesn't have to -- as long as the hash doesn't change I think it's fine.