@haampie 2023-04-20
I benchmarked it once more with `spack solve --timers cp2k+sirius` on some hpc system:

Before:
```
    setup         20.348s
    load           0.037s
    ground         4.797s
    solve         11.239s
    total         37.111s
```

After:
```
    setup         14.898s
    load           0.035s
    ground         4.732s
    solve         11.229s
    total         32.122s
```